Don't rebuild the GitLab docker image to set an environment variable
Before this patch we used a workaround taken directly from
https://gitlab.com/gitlab-org/omnibus-gitlab/-/issues/2837#note_109602925,
however the workaround is far more involved than it needs to be.

Instead of setting `ENV` in a Dockerfile, we can simply set an
environment variable on the CLI with the same effect, saving users the
time it takes to build a docker image, as well as the maintenance
overhead of running a custom build script for updates.