fix copy commands in setup
`cp -r` command will create a new directory if the destination
directory does not exist, but create sub-directory under it if
it exists. This results a nested scripts/scripts folder to be
created which is not intended. This causes problem for docker
volume binding.

remove the tailing folder names to get rid of this issue.

Signed-off-by: Kejia Hu <kejia.hu@codethink.co.uk>