Fix Docker-compose installers '/scripts' volume structure. (#428)
The Docker-compose-based installers overwrite the COMMANDs in the
container images of the Engine and SDLC components with some scripts
which handle some setup tasks before starting the actual services.

Part of the setup stepts these scripts do is attempting to run a
`gen-config.sh` script they assume is located in `/scripts`, but the
various pre-Docker-compose scripts accidentally copy `gen-config.sh`
into a subdirectory also called '/scripts'. (thus, inside the containers,
the actual path would be '/scripts/scripts/gen-config.sh')

This patch ensures that all of the scripts required by the Engine and
SDLC containers are located under the single '/scripts' directory as
was originally intended.