Run tests for showcases
Adds workflow to test changes to showcases twice daily, on PRs or commit on master.

Showcase tests consist of a roundtrip compilation test with latest legend versions to enforce compilation and formatting. Does not yet allow for comments.

The paths in the poms and the test file are different due to the
location that workflows vs tests checkout the code.

Showcase code moved to not clog up directory structure.